feat(engine): trigger compensations in the right order
When a process instance enters a compensation intermediate throw or end event,
it triggers the compensation of the current scope and invokes
 the compensation handlers of completed activities.

Next to the current scope, it should trigger the compensation
for the child scopes recursively (i.e. propagate from top to bottom).
As a result, it should invoke the compensation handlers in completed embedded subprocesses.
If a subprocess is not completed (i.e. still active or terminated),
the compensation handlers within the subprocess should not be invoked.