fix: copy column families from main
This allows us to easier backport changes from main related to column
families.

To make this change I also had to change the usage of some column
families that have been renamed for deprecation.

For the rest this change is safe to make, because it's no problem to
have unused column families.